requirements.txt
# Core packages
flask>=2.0.0
streamlit>=1.24.0
requests>=2.28.0
python-dotenv>=0.19.0
Pillow>=9.0.0

# Marketing agent specific packages
beautifulsoup4>=4.9.3
selenium>=4.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
nltk>=3.6.0
textblob>=0.15.3
yfinance>=0.1.63
matplotlib>=3.4.0
seaborn>=0.11.0
chromadb>=0.3.0
langchain>=0.0.150
openai>=0.27.0
python-dateutil>=2.8.2
tqdm>=4.62.0
google-search-results>=2.4.2

# Additional dependencies
urllib3>=1.26.0
certifi>=2021.5.30
idna>=3.2
charset-normalizer>=2.0.0
lxml>=4.9.0
soupsieve>=2.0
webdriver-manager>=3.5.0
joblib>=1.0.2
threadpoolctl>=2.0.2
scipy>=1.7.0
